Add files via upload
This is a test to use Mask R-CNN. The approach here is to use R-CNN to detect cars and then use intersection over union method to find the empty spots. Now this code show that R-CNN is bad at detecting cars especially when the cars are small. Therefore, it does not make sense to even proceed to step two and the team will approach the image segmentation manually.